build: Add support for selective LTO with QCLD
This patch adjusts the current SDCLANG support
with selective LTO and QCLD support.

QCLD is Qualcomm's proprietary linker which
includes major improvements for LTO.

With the current implementation LTO and QCLD
can be enabled together and it is impossible
to enable one without the other.
This is because that's the most benefitial
configuration.

Since SDCLANG bundles QCLD, using
QCLD without enabling SDCLANG globally
is blocked.

On static libraries the LTO flags
don't get applied, the control file
living in device/qcom/common/sdclang/
has special handling for this.

Global builds with this are not possible
at the current state, since QCLD misses
certain functionalities gold offers and
some targets depend on these functionalities.
To enable this all a target has to do
is setting LOCAL_SDCLANG_LTO to true in its makefile.

Flags which are incompatible with QCLD get
filtered by including the variable SDCLANG_LTO_DEFS .
Check device/qcom/common for more details.

Contributors:
@TheCrazyLex
@xboxfanj

Change-Id: Ib189d3b983a3fd78ef5bd5ab9fc1ff1351a506f4
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: BijiOnta <sengolcaduk32@gmail.com>